requirements.txt
faiss-cpu==1.9.0
rank-bm25==0.2.2
numpy==2.1.3
python-dotenv==1.0.1
google-generativeai==0.8.4
